pyproject.toml
[project]
name = "memory-bank-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.0.2",
    "python-dotenv>=1.2.1",
    "upstash-redis>=1.5.0",
]
